The "no free pages left" check is now executed during the compare-exchange loop as well
Omitting that check during the loop wasn't strictly correct, and I find strict correctness
important in what's supposed to be a secure malloc implementation.

Performance impact should be minimal, as the loop is unlikely to run even for one iteration.
The case the loop protects against is rare in practice.